Moved now "common" cosmos things to common-x files. Deleted old cosmosdb prefixed files.
