+++ pkg/services/cosmosdb/common-bind.go
