+++ pkg/services/cosmosdb/common-update.go
