--- pkg/services/cosmosdb/cosmosdb-types.go
@@ -1,27 +1,0 @@
-package cosmosdb
-
-import "github.com/Azure/open-service-broker-azure/pkg/service"
-
-// cosmosCredentials encapsulates CosmosDB-specific details for connecting via
-// a variety of APIs. This excludes MongoDB.
-type cosmosCredentials struct {
-	URI                     string `json:"uri,omitempty"`
-	PrimaryConnectionString string `json:"primaryConnectionString,omitempty"`
-	PrimaryKey              string `json:"primaryKey,omitempty"`
-}
-
-func (c *cosmosAccountManager) SplitProvisioningParameters(
-	cpp service.CombinedProvisioningParameters,
-) (
-	service.ProvisioningParameters,
-	service.SecureProvisioningParameters,
-	error,
-) {
-	return nil, nil, nil
-}
-
-func (c *cosmosAccountManager) SplitBindingParameters(
-	params service.CombinedBindingParameters,
-) (service.BindingParameters, service.SecureBindingParameters, error) {
-	return nil, nil, nil
-}
